+++ voice/index.ts
@@ -1,4 +1,3 @@
 export * from "./types";
 export * from "./VoiceConnection";
-export * from "./VoicePeerConnection";
 export * from "./AudioPlayer";